+++ src/main/java/org/stellar/sdk/Transaction.java
@@ -19,8 +19,6 @@ import static com.google.common.base.Preconditions.checkNotNull;
  * Represents <a href="https://www.stellar.org/developers/learn/concepts/transactions.html" target="_blank">Transaction</a> in Stellar network.
  */
 public class Transaction {
-  private static final int BASE_FEE = 100;
-
   private final int mFee;
   private final KeyPair mSourceAccount;
   private final long mSequenceNumber;
@@ -219,13 +217,9 @@ public class Transaction {
   }
 
   /**
-   * Generates TransactionEnvelope XDR object. Transaction need to have at least one signature.
+   * Generates TransactionEnvelope XDR object.
    */
   public org.stellar.sdk.xdr.TransactionEnvelope toEnvelopeXdr() {
-    if (mSignatures.size() == 0) {
-      throw new NotEnoughSignaturesException("Transaction must be signed by at least one signer. Use transaction.sign().");
-    }
-
     org.stellar.sdk.xdr.TransactionEnvelope xdr = new org.stellar.sdk.xdr.TransactionEnvelope();
     org.stellar.sdk.xdr.Transaction transaction = this.toXdr();
     xdr.setTx(transaction);
@@ -257,13 +251,24 @@ public class Transaction {
    * Builds a new Transaction object.
    */
   public static class Builder {
+    private static final int BASE_FEE = 100;
     private final TransactionBuilderAccount mSourceAccount;
     private Memo mMemo;
     private TimeBounds mTimeBounds;
     List<Operation> mOperations;
     private boolean timeoutSet;
+    private static Integer defaultOperationFee;
+    private Integer operationFee;
 
     public static final long TIMEOUT_INFINITE = 0;
+
+    public static void setDefaultOperationFee(int opFee) {
+      if (opFee < BASE_FEE) {
+        throw new IllegalArgumentException("DefaultOperationFee cannot be smaller than the BASE_FEE (" + BASE_FEE + "): " + opFee);
+      }
+
+      defaultOperationFee = opFee;
+    }
 
     /**
      * Construct a new transaction builder.
@@ -275,6 +280,7 @@ public class Transaction {
       checkNotNull(sourceAccount, "sourceAccount cannot be null");
       mSourceAccount = sourceAccount;
       mOperations = Collections.synchronizedList(new ArrayList<Operation>());
+      operationFee = defaultOperationFee;
     }
 
     public int getOperationsCount() {
@@ -361,6 +367,15 @@ public class Transaction {
       return this;
     }
 
+    public Builder setOperationFee(int operationFee) {
+      if (operationFee < BASE_FEE) {
+        throw new IllegalArgumentException("OperationFee cannot be smaller than the BASE_FEE (" + BASE_FEE + "): " + operationFee);
+      }
+
+      this.operationFee = operationFee;
+      return this;
+    }
+
     /**
      * Builds a transaction. It will increment sequence number of the source account.
      */
@@ -370,9 +385,14 @@ public class Transaction {
         throw new RuntimeException("TimeBounds has to be set or you must call setTimeout(TIMEOUT_INFINITE).");
       }
 
+      if (operationFee == null) {
+        System.out.println("[TransactionBuilder] The `operationFee` parameter of `TransactionBuilder` is required. Setting to BASE_FEE=" + BASE_FEE + ". Future versions of this library will error if not provided.");
+        operationFee = BASE_FEE;
+      }
+
       Operation[] operations = new Operation[mOperations.size()];
       operations = mOperations.toArray(operations);
-      Transaction transaction = new Transaction(mSourceAccount.getKeypair(), operations.length * BASE_FEE, mSourceAccount.getIncrementedSequenceNumber(), operations, mMemo, mTimeBounds);
+      Transaction transaction = new Transaction(mSourceAccount.getKeypair(), operations.length * operationFee, mSourceAccount.getIncrementedSequenceNumber(), operations, mMemo, mTimeBounds);
       // Increment sequence number when there were no exceptions when creating a transaction
       mSourceAccount.incrementSequenceNumber();
       return transaction;