+++ src/main/java/org/stellar/sdk/responses/OfferResponse.java
@@ -26,10 +26,14 @@ public class OfferResponse extends Response {
   private final String amount;
   @SerializedName("price")
   private final String price;
+  @SerializedName("last_modified_ledger")
+  private final Integer lastModifiedLedger;
+  @SerializedName("last_modified_time")
+  private final String lastModifiedTime;
   @SerializedName("_links")
   private final Links links;
 
-  OfferResponse(Long id, String pagingToken, KeyPair seller, Asset selling, Asset buying, String amount, String price, Links links) {
+  public OfferResponse(Long id, String pagingToken, KeyPair seller, Asset selling, Asset buying, String amount, String price, Integer lastModifiedLedger, String lastModifiedTime, Links links) {
     this.id = id;
     this.pagingToken = pagingToken;
     this.seller = seller;
@@ -37,6 +41,8 @@ public class OfferResponse extends Response {
     this.buying = buying;
     this.amount = amount;
     this.price = price;
+    this.lastModifiedLedger = lastModifiedLedger;
+    this.lastModifiedTime = lastModifiedTime;
     this.links = links;
   }
 
@@ -66,6 +72,15 @@ public class OfferResponse extends Response {
 
   public String getPrice() {
     return price;
+  }
+
+  public Integer getLastModifiedLedger() {
+    return lastModifiedLedger;
+  }
+
+  // Can be null if ledger adding an offer has not been ingested yet.
+  public String getLastModifiedTime() {
+    return lastModifiedTime;
   }
 
   public Links getLinks() {